Wip on eagreement test
