+++ build.gradle
@@ -130,7 +130,8 @@ task genJaxb {
 // end::wsdl[]
 
 dependencies {
-	ktlint "com.github.shyiko:ktlint:0.22.0"
+	testImplementation 'org.junit.jupiter:junit-jupiter:5.8.1'
+    ktlint "com.github.shyiko:ktlint:0.22.0"
 
 	compile("com.taktik.boot:spring-boot-starter-micrometer:2.1-90-gcc5c980239")
 	compile('org.springframework.boot:spring-boot-starter-jetty')
