+++ src/test/kotlin/org/taktik/freehealth/middleware/web/controllers/AgreementControllerTest.kt
@@ -45,6 +45,4 @@ class AgreementControllerTest : EhealthTest() {
         HttpMethod.POST, HttpEntity<Void>(createHeaders(null, null, keystoreId, tokenId, passPhrase)), String::class.java, passPhrase).body
 
     }
-
-
 }
